third_party/blink/web_tests/external/wpt/navigation-api/navigation-activation/activation-initial-about-blank-expected.txt

This is a testharness.js-based test.
[FAIL] navigation.activation interaction with initial about:blank
  assert_equals: expected null but got object "[object NavigationHistoryEntry]"
Harness: the test ran to completion.
